+++ lib/features/canvas/pages/canvas_page.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:scribble/scribble.dart';
 
-import '../features/canvas/models/tool_mode.dart';
-import '../features/canvas/notifiers/custom_scribble_notifier.dart';
-import '../features/canvas/notifiers/scribble_notifier_x.dart';
-import '../features/canvas/widgets/background_placeholder.dart';
-import '../features/canvas/widgets/canvas_info.dart';
-import '../features/canvas/widgets/canvas_toolbar.dart';
-import '../features/notes/models/note.dart';
+import '../../notes/models/note.dart';
+import '../models/tool_mode.dart';
+import '../notifiers/custom_scribble_notifier.dart';
+import '../notifiers/scribble_notifier_x.dart';
+import '../widgets/background_placeholder.dart';
+import '../widgets/canvas_info.dart';
+import '../widgets/canvas_toolbar.dart';
 
 class CanvasPage extends StatefulWidget {
   const CanvasPage({
